+++ lib/gen/protos/protos.pbjson.dart
@@ -1,0 +1,35 @@
+///
+//  Generated code. Do not modify.
+//  source: protos/protos.proto
+//
+// @dart = 2.3
+// ignore_for_file: camel_case_types,non_constant_identifier_names,library_prefixes,unused_import,unused_shown_name,return_of_invalid_type
+
+const BarcodeFormat$json = const {
+  '1': 'BarcodeFormat',
+  '2': const [
+    const {'1': 'aztec', '2': 0},
+    const {'1': 'code39', '2': 1},
+    const {'1': 'code93', '2': 2},
+    const {'1': 'ean8', '2': 3},
+    const {'1': 'ean13', '2': 4},
+    const {'1': 'code128', '2': 5},
+    const {'1': 'dataMatrix', '2': 6},
+    const {'1': 'qr', '2': 7},
+    const {'1': 'interleaved2of5', '2': 8},
+    const {'1': 'upce', '2': 9},
+    const {'1': 'pdf417', '2': 10},
+  ],
+};
+
+const Configuration$json = const {
+  '1': 'Configuration',
+  '2': const [
+    const {'1': 'cancelText', '3': 1, '4': 1, '5': 9, '10': 'cancelText'},
+    const {'1': 'flashOnText', '3': 2, '4': 1, '5': 9, '10': 'flashOnText'},
+    const {'1': 'flashOffText', '3': 3, '4': 1, '5': 9, '10': 'flashOffText'},
+    const {'1': 'restrictFormat', '3': 4, '4': 3, '5': 14, '6': '.BarcodeFormat', '10': 'restrictFormat'},
+    const {'1': 'useCamera', '3': 5, '4': 1, '5': 5, '10': 'useCamera'},
+  ],
+};
+